// webapp/src/app/backend/model/checkout.ts
/**
 * Shopping Advanced
 * API for mobile apps
 *
 * The version of the OpenAPI document: \'0.0.1\'
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Checkout { 
    readonly time?: string;
    trip: number;
    item: number;
    count?: number;
}